+++ pueue/src/client/query/mod.rs
@@ -189,6 +189,7 @@ pub fn apply_query(query: &str, group: &Option<String>) -> Result<QueryResult> {
             Rule::column_selection => column_selection::apply(section, &mut query_result)?,
             Rule::datetime_filter => filters::datetime(section, &mut query_result)?,
             Rule::label_filter => filters::label(section, &mut query_result)?,
+            Rule::command_filter => filters::command(section, &mut query_result)?,
             Rule::status_filter => filters::status(section, &mut query_result)?,
             Rule::order_by_condition => order_by::order_by(section, &mut query_result)?,
             Rule::limit_condition => limit::limit(section, &mut query_result)?,